P1/P1-ws-server/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
dj-database-url==2.2.0
Django==4.2.13
django-modern-rpc==1.0.3
djangorestframework==3.15.2
Faker==30.8.0
flake8==6.1.0
google==3.0.0
grpc-django==1.0.0
grpcio==1.67.1
grpcio-tools==1.67.1
gunicorn==20.1.0
idna==3.10
protobuf==5.28.3
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==0.21.0
requests==2.32.3
#si2_models @ git+ssh://git@github.com/rmarabini/SI2Django.git@310debc06eb73cb79a02dfb808f607109367ad49#subdirectory=VotingProj_v2/Common
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.7.0
djangorestframework-simplejwt==4.4.0
django-model-utils==4.0.0